pyproject.toml
[project]
name = "agents-proj-15"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "openai-agents",
  "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "src"}
